Regression test portability: batch of bugs (volume 1)
Fix various regression test portability in tests/bugs.

bug-1004744.t:
- Slower systems really requires an increased REBALANCE_TIMEOUT in include.rc

bug-1023974.t:
- use the -p option to mkdir before the path for portability sake.

bug-1051896.t and bug-847622.t:
- skip ACL test for NetBSD as it has no POSIX 1e ACL support.

bugs-1053579.t:
- Override the amount of secondary groups depending of the maximum
  the system supports.
- Specify seq(1) format to avoid having trailing .00
- Use more portable su -m USER -c COMMAND (also in bug-884597.t)
- NetBSD does not fail high GID with NFS

bug-1058663.c:
- Depending on architecture, NetBSD can have SIGSEGV instead of SIGBUS

bug-762989.t and bug-867253.t:
- For non Linux systems, skip tests on features that are Linux-specific

bug-765473.t;
- Fix a {1} that bash is supposed to expand in 1 but does not, replace by 1.
- Sync volume to make sure it imediatly knows a fd got bad

BUG: 1129939
Reviewed-on: http://review.gluster.org/9009
Tested-by: Gluster Build System <jenkins@build.gluster.com>

+++ tests/bugs/bug-762989.t
@@ -4,6 +4,14 @@
 
 cleanup;
 
+# Skip the entire test if ip_local_reserved_ports does not exist
+if [ ! -f /proc/sys/net/ipv4/ip_local_reserved_ports ] ; then
+    echo "Skip test on /proc/sys/net/ipv4/ip_local_reserved_ports, "\
+         "which does not exists on this system" >&2
+    SKIP_TESTS
+    exit 0
+fi
+
 ## reserve port 1023
 older_ports=$(cat /proc/sys/net/ipv4/ip_local_reserved_ports);
 echo "1023" > /proc/sys/net/ipv4/ip_local_reserved_ports;
